# Convert performance tests to use BenchmarkDotNet