--- a/FastMember_Net40.Signed.Tests/packages.config
+++ b/FastMember_Net40.Signed.Tests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="BenchmarkDotNet" version="0.9.1" targetFramework="net45" />
   <package id="NUnit" version="3.0.0" targetFramework="net45" />
 </packages>