--- a/FastMember_Net40.Tests/packages.config
+++ b/FastMember_Net40.Tests/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NUnit" version="3.0.0" targetFramework="net40" />
+  <package id="BenchmarkDotNet" version="0.9.1" targetFramework="net4" />
+  <package id="NUnit" version="3.0.0" targetFramework="net4" />
 </packages>