@dotnetchris 2016-03-15
@mattwarren i really like the scale number, makes it far more understandable.