@mattwarren 2016-03-16
@dotnetchris yeah they are helpful, it was actually an accidental feature. 

We found that people using BenchmarkDotNet were calculating the `scaled` value themselves and manually putting it into the results, so we figured it would be easier if we made is a built-in option.